// frontend/src/types/types.ts
export class ReviewResponse {
    id: number;
    author: string;
    message: string;
    rating: number;
    date: any;
    constructor(data: { id: number; author: string; message: string;rating:number;date:any }) {
        const { id, author, message,rating,date } = data;
        this.id = id;
        this.author = author;
        this.message = message;
        this.rating = rating;
        this.date = date
      }
    

}

export interface ReviewRequest {
    perfumeId: number | string;
    author: string;
    message: string;
    rating: number;
}

export interface ReviewError {
    authorError: string;
    messageError: string;
    ratingError: string;
}